requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
pydantic==2.9.1
python-dotenv==0.20.0
SQLAlchemy==2.0.30
PyMySQL==1.1.1
pandas==2.2.2
requests==2.32.3
